+++ Classes/Collection/Collection.php
@@ -14,11 +14,12 @@ namespace Brotkrueml\FeedGenerator\Collection;
 use Brotkrueml\FeedGenerator\Contract\AttachmentInterface;
 use Brotkrueml\FeedGenerator\Contract\AuthorInterface;
 use Brotkrueml\FeedGenerator\Contract\CategoryInterface;
+use Brotkrueml\FeedGenerator\Contract\ExtensionElementInterface;
 use Brotkrueml\FeedGenerator\Contract\ItemInterface;
 use Traversable;
 
 /**
- * @template T of AttachmentInterface|AuthorInterface|CategoryInterface|ItemInterface
+ * @template T of AttachmentInterface|AuthorInterface|CategoryInterface|ExtensionElementInterface|ItemInterface
  * @implements \IteratorAggregate<T>
  */
 final class Collection implements \IteratorAggregate
@@ -32,7 +33,7 @@ final class Collection implements \IteratorAggregate
      * @param T $items
      * @return self<T>
      */
-    public function add(AttachmentInterface|AuthorInterface|CategoryInterface|ItemInterface ...$items): self
+    public function add(AttachmentInterface|AuthorInterface|CategoryInterface|ExtensionElementInterface|ItemInterface ...$items): self
     {
         if ($items === []) {
             return $this;
@@ -53,7 +54,7 @@ final class Collection implements \IteratorAggregate
     /**
      * @return T
      */
-    public function get(int $index): AttachmentInterface|AuthorInterface|CategoryInterface|ItemInterface
+    public function get(int $index): AttachmentInterface|AuthorInterface|CategoryInterface|ExtensionElementInterface|ItemInterface
     {
         return $this->items[$index] ?? throw IndexNotFoundException::forIndex($index);
     }