+++ Classes/Configuration/ExtensionRegistryInterface.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the "feed_generator" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+namespace Brotkrueml\FeedGenerator\Configuration;
+
+use Brotkrueml\FeedGenerator\Contract\ExtensionElementInterface;
+use Brotkrueml\FeedGenerator\Contract\ExtensionInterface;
+
+/**
+ * @internal
+ */
+interface ExtensionRegistryInterface
+{
+    public function getExtensionForElement(ExtensionElementInterface $element): ExtensionInterface;
+
+    /**
+     * @return iterable<ExtensionInterface>
+     */
+    public function getAllExtensions(): iterable;
+}