+++ Classes/Entity/AbstractFeed.php
@@ -14,6 +14,7 @@ namespace Brotkrueml\FeedGenerator\Entity;
 use Brotkrueml\FeedGenerator\Collection\Collection;
 use Brotkrueml\FeedGenerator\Contract\AuthorInterface;
 use Brotkrueml\FeedGenerator\Contract\CategoryInterface;
+use Brotkrueml\FeedGenerator\Contract\ExtensionElementInterface;
 use Brotkrueml\FeedGenerator\Contract\FeedInterface;
 use Brotkrueml\FeedGenerator\Contract\ImageInterface;
 use Brotkrueml\FeedGenerator\Contract\ItemInterface;
@@ -102,5 +103,13 @@ abstract class AbstractFeed implements FeedInterface
     public function getStyleSheet(): string
     {
         return '';
+    }
+
+    /**
+     * @return Collection<ExtensionElementInterface>
+     */
+    public function getExtensionElements(): Collection
+    {
+        return new Collection();
     }
 }