+++ Tests/Fixtures/Renderer/Atom/FullFeed.php
@@ -115,4 +115,9 @@ final class FullFeed implements FeedInterface
     {
         return '';
     }
+
+    public function getExtensionElements(): Collection
+    {
+        return new Collection();
+    }
 }